+++ src/structures/Friend.ts
@@ -139,6 +139,18 @@ class Friend extends User {
   public async getMutualFriends() {
     return this.client.getMutualFriends(this.id);
   }
+
+  /**
+   * Checks whether this friend owns a specific offer
+   * @param friend The id or display name of the friend
+   * @param offerId The offer id
+   * @throws {OfferNotFoundError} The offer does not exist or is not in the current storefront catalog
+   * @throws {FriendNotFoundError} The user does not exist or is not friends with the client
+   * @throws {EpicgamesAPIError}
+   */
+  public async checkOfferOwnership(offerId: string) {
+    return this.client.checkFriendOfferOwnership(this.id, offerId);
+  }
 }
 
 export default Friend;